refactor: replaced old component with new one and fixed minor layout changes
