+++ src/Components/ProposalInfoCard/index.ts
